Added and Tested More Database Functionality

+++ app/src/main/java/com/seannajera/golfiq/model/persistenceDb/entity/MarkerLocation.kt
@@ -5,4 +5,4 @@ import android.arch.persistence.room.PrimaryKey
 import java.util.*
 
 @Entity()
-class MarkerLocation(@PrimaryKey var id: Int, var lat: Double, var lon: Double, var createdAt: Date)
+class MarkerLocation(@PrimaryKey var id: Long, var lat: Double, var lon: Double, var createdAt: Date)
