+++ app/src/main/java/com/seannajera/golfiq/model/persistenceDb/entity/UserLocation.kt
@@ -5,4 +5,4 @@ import android.arch.persistence.room.PrimaryKey
 import java.util.*
 
 @Entity()
-class UserLocation(@PrimaryKey var id: Int, var lat: Double, var lon: Double, var errorMargin: Double, var createdAt: Date)
+class UserLocation(@PrimaryKey var id: Long, var lat: Double, var lon: Double, var errorMargin: Double, var createdAt: Date)
